Add files via upload
Monkey eat peaches. First day, it ate half of the total peaches, and plus one peach; second day, it ate the half of the left amount, and one more peach. So on....
On the morning of the fifth day, only one peach left. Ask how many peaches are there on the first day?